// src/modules/transfer/domain/constants/additional-data-key.enum.ts
export enum AdditionalDataKey {
  RESPONSE_CODE = 'RESPONSE_CODE',
  RESPONSE_MESSAGE = 'RESPONSE_MESSAGE', //TODO
  DOCUMENT_NUMBER_KEY_RESOLUTION = 'DOCUMENT_NUMBER_KEY_RESOLUTION',
  OWNER_NAME_KEY_RESOLUTION = 'OWNER_NAME_KEY_RESOLUTION',
  ACCOUNT_NUMBER_KEY_RESOLUTION = 'ACCOUNT_NUMBER_KEY_RESOLUTION'
}